+++ config/machines/lumi/container/load_aqua_lumi.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-AQUA_container="/project/project_465000454/containers/aqua/aqua-v0.3.sif"
+AQUA_container="/project/project_465000454/containers/aqua/aqua-v0.4.sif"
 FDB5_CONFIG_FILE="/scratch/project_465000454/igonzalez/fdb-long/config.yaml"
 GSV_WEIGHTS_PATH="/scratch/project_465000454/igonzalez/gsv_weights/"
 GRID_DEFINITION_PATH="/scratch/project_465000454/igonzalez/grid_definitions"
@@ -13,6 +13,7 @@ singularity shell \
     --env PYTHONPATH=/opt/conda/lib/python3.10/site-packages \
     --env ESMFMKFILE=/opt/conda/lib/esmf.mk \
     --bind /pfs/lustrep3/ \
+    --bind /projappl/ \
     --bind /scratch/ \
     $AQUA_container
 # Run this script in LUMI in VSCode