+++ config/machines/lumi/container/slurm_job_using_container.sh
@@ -9,7 +9,7 @@
 #SBATCH --error=aqua_slurm.err
 #SBATCH -p debug    #change the partition
 
-AQUA_container=/project/project_465000454/containers/aqua/aqua-v0.3.sif
+AQUA_container=/project/project_465000454/containers/aqua/aqua-v0.4.sif
 FDB5_CONFIG_FILE=/scratch/project_465000454/igonzalez/fdb-long/config.yaml #Change it to your simulation
 GSV_WEIGHTS_PATH=/scratch/project_465000454/igonzalez/gsv_weights/
 GRID_DEFINITION_PATH=/scratch/project_465000454/igonzalez/grid_definitions
@@ -22,6 +22,7 @@ singularity exec  \
     --env PYTHONPATH=/opt/conda/lib/python3.10/site-packages \
     --env ESMFMKFILE=/opt/conda/lib/esmf.mk  \
     --bind /pfs/lustrep3/ \
+    --bind /projappl/ \
     --bind /scratch/ \
     $AQUA_container \
     bash -c \