+++ diagnostics/ocean3d/cli/ocean3d_cli.py
@@ -1,16 +1,12 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-import sys
+
 import argparse
 import os
 import sys
 
 from aqua import Reader
 from aqua.util import load_yaml, get_arg, create_folder
-from aqua.exceptions import NoObservationError
-
-# This is needed if loading from the cli directory
-sys.path.insert(0, '../../..')
-sys.path.insert(0, '../..')
 
 from ocean3d import plot_stratification
 from ocean3d import plot_spatial_mld_clim
@@ -20,6 +16,9 @@ from ocean3d import time_series_multilevs
 from ocean3d import multilevel_t_s_trend_plot
 from ocean3d import zonal_mean_trend_plot
 
+from aqua.util import find_vert_coord
+from aqua.logger import log_configure
+
 
 def parse_arguments(args):
     """Parse command line arguments"""
@@ -28,6 +27,8 @@ def parse_arguments(args):
 
     parser.add_argument('--config', type=str,
                         help='yaml configuration file')
+    parser.add_argument('-l', '--loglevel', type=str,
+                        help='log level [default: WARNING]')
 
     # This arguments will override the configuration file is provided
     parser.add_argument('--model', type=str, help='Model name')
@@ -39,102 +40,184 @@ def parse_arguments(args):
     return parser.parse_args(args)
 
 
-if __name__ == '__main__':
+def ocean3d_diags(data, region=None,
+                  latS: float = None,
+                  latN: float = None,
+                  lonW: float = None,
+                  lonE: float = None,
+                  output_dir: str = None,
+                  loglevel: str = 'WARNING'):
 
-    print("Running ocean3d diagnostic...")
+    logger = log_configure(log_name='Ocean3D Diagnostic', log_level=loglevel)
+
+    logger.debug("Evaluating Hovmoller plots")
+    hovmoller_lev_time_plot(data=data,
+                            lonE=lonE, lonW=lonW, latS=latS, latN=latN, region=region,
+                            anomaly=False, standardise=False,
+                            output=True, output_dir=output_dir)
+    hovmoller_lev_time_plot(data=data,
+                            lonE=lonE, lonW=lonW, latS=latS, latN=latN, region=region,
+                            anomaly=False, standardise=True,
+                            output=True, output_dir=output_dir)
+    hovmoller_lev_time_plot(data=data,
+                            lonE=lonE, lonW=lonW, latS=latS, latN=latN,
+                            region=region, anomaly=True, anomaly_ref="t0", standardise=False,
+                            output=True, output_dir=output_dir)
+    hovmoller_lev_time_plot(data=data,
+                            lonE=lonE, lonW=lonW, latS=latS, latN=latN, region=region,
+                            anomaly=True, anomaly_ref="tmean", standardise=False,
+                            output=True, output_dir=output_dir)
+    hovmoller_lev_time_plot(data=data,
+                            lonE=lonE, lonW=lonW, latS=latS, latN=latN, region=region,
+                            anomaly=True, anomaly_ref="t0", standardise=True,
+                            output=True, output_dir=output_dir)
+    hovmoller_lev_time_plot(data=data,
+                            lonE=lonE, lonW=lonW, latS=latS, latN=latN, region=region,
+                            anomaly=True, anomaly_ref="tmean", standardise=True,
+                            output=True, output_dir=output_dir)
+
+    logger.debug("Evaluating time series multilevels")
+    time_series_multilevs(data=data,
+                          lonE=lonE, lonW=lonW, latS=latS, latN=latN, region=region,
+                          anomaly=False, standardise=False, customise_level=False, levels=list,
+                          output=True, output_dir=output_dir)
+    time_series_multilevs(data=data,
+                          lonE=lonE, lonW=lonW, latS=latS, latN=latN, region=region,
+                          anomaly=True, standardise=False, anomaly_ref="tmean", customise_level=False, levels=list,
+                          output=True, output_dir=output_dir)
+    time_series_multilevs(data=data,
+                          lonE=lonE, lonW=lonW, latS=latS, latN=latN, region=region,
+                          anomaly=True, standardise=False, anomaly_ref="t0", customise_level=False, levels=list,
+                          output=True,  output_dir=output_dir)
+    time_series_multilevs(data=data,
+                          lonE=lonE, lonW=lonW, latS=latS, latN=latN, region=region,
+                          anomaly=True, standardise=True, anomaly_ref="tmean", customise_level=False, levels=list,
+                          output=True, output_dir=output_dir)
+    time_series_multilevs(data=data,
+                          lonE=lonE, lonW=lonW, latS=latS, latN=latN, region=region,
+                          anomaly=True, standardise=True, anomaly_ref="t0", customise_level=False, levels=list,
+                          output=True, output_dir=output_dir)
+
+    logger.debug("Evaluating multilevel_t_s_trend_plot")
+    multilevel_t_s_trend_plot(data=data,
+                              lonE=lonE, lonW=lonW, latS=latS, latN=latN, region=region,
+                              customise_level=False, levels=None,
+                              output=True, output_dir=output_dir)
+
+    logger.debug("Evaluating zonal_mean_trend_plot")
+    zonal_mean_trend_plot(data=data,
+                          lonE=lonE, lonW=lonW, latS=latS, latN=latN, region=region,
+                          output=True, output_dir=output_dir)
+
+    for time in range(1, 18):  # 1 to 12 is the months, then each number directs the seasonals and the yearly climatology
+        logger.debug("Evaluating plot_stratification, time: %s", time)
+        plot_stratification(mod_data=data,
+                            lonE=lonE, lonW=lonW, latS=latS, latN=latN, region=region,
+                            time=time,
+                            output=True, output_dir=output_dir)
+        # plot_spatial_mld_clim(mod_data= data, lonE=lonE, lonW=lonW, latS=latS, latN=latN, region=region, time = time, overlap= True,output= True, output_dir= output_dir)
+
+
+def get_value_with_default(dictionary, key, default_value):
+    try:
+        return dictionary[key]
+    except KeyError:
+        return default_value
+
+
+if __name__ == '__main__':
     args = parse_arguments(sys.argv[1:])
+
+    loglevel = get_arg(args, 'loglevel', 'WARNING')
+    logger = log_configure(log_name='Ocean3D CLI', log_level=loglevel)
+
+    # change the current directory to the one of the CLI so that relative path works
+    abspath = os.path.abspath(__file__)
+    dname = os.path.dirname(abspath)
+    if os.getcwd() != dname:
+        os.chdir(dname)
+        logger.info(f'Moving from current directory to {dname} to run!')
+
+    logger.info("Running ocean3d diagnostic...")
 
     # Read configuration file
     file = get_arg(args, 'config', 'config.yaml')
-    print('Reading configuration yaml file..')
+    logger.info('Reading configuration yaml file..')
+
     ocean3d_config = load_yaml(file)
+
+    logger.debug(f"Configuration file: {ocean3d_config}")
 
     model = get_arg(args, 'model', ocean3d_config['model'])
     exp = get_arg(args, 'exp', ocean3d_config['exp'])
     source = get_arg(args, 'source', ocean3d_config['source'])
     outputdir = get_arg(args, 'outputdir', ocean3d_config['outputdir'])
 
-    create_folder(outputdir)
+    custom_regions = get_value_with_default(ocean3d_config,
+                                            "custom_region", [])
+    predefined_regions = get_value_with_default(ocean3d_config,
+                                                "predefined_regions", [])
 
-    print(f"Reader selecting for model={model}, exp={exp}, source={source}")
+    logger.debug(f"custom_region: {custom_regions}")
+    logger.debug(f"predefined_regions: {predefined_regions}")
+
+    create_folder(outputdir, loglevel=loglevel)
+
+    logger.info(f"Reader selecting for model={model}, exp={exp}, source={source}")
     try:
-        reader = Reader(model, exp, source, fix=True)
+        reader = Reader(model=model, exp=exp, source=source,
+                        fix=True, loglevel=loglevel)
+        data = reader.retrieve()
+
+        vertical_coord = find_vert_coord(data)[0]
+        data = data.rename({vertical_coord: "lev"})
     except KeyError:
         # NOTE: This should be a proper NoDataError
-        print("NoDataError: No data available")
+        logger.error("NoDataError: No data available")
+        sys.exit(0)
+    except Exception as e:
+        logger.error(f"Error: {e}")
+        logger.error("This could a bug, please report it to the developers.")
         sys.exit(0)
 
-    data = reader.retrieve()
+    try:
+        if custom_regions:
+            logger.info("Analysing custom regions")
+            custom_regions = ocean3d_config["custom_region"] ### add fix if not present
+            custom_region_dict = {}
+            for custom_region in custom_regions:
+                for coord in custom_region:
+                    custom_region_dict.update(coord)
+                lonE = custom_region_dict["lonE"]
+                lonW = custom_region_dict["lonW"]
+                latS = custom_region_dict["latS"]
+                latN = custom_region_dict["latN"]
 
-    # HACK: FESOM data has nz1 as the vertical dimension
-    #       Check issue #531 for more details
-    if model == 'FESOM':
-        data = data.rename({"nz1": "lev"})
+                logger.debug("lonE: %s, lonW: %s, latS: %s, latN: %s",
+                             lonE, lonW, latS, latN)
+
+                ocean3d_diags(data,
+                              region=None, latS=latS, latN=latN, lonW=lonW, lonE=lonE,
+                              output_dir=outputdir, loglevel=loglevel)
+    except AttributeError:
+        logger.error("NoDataError: so or ocpt not found in the Dataset.")
+        logger.critical("Not producting ocean diagnostics for custom regions.")
+    except Exception as e:
+        logger.error(f"Error: {e}, not producting ocean diagnostics for custom regions.")
+        logger.critical("This could a bug, please report it to the developers.")
 
     try:
-        hovmoller_lev_time_plot(data=data, region="Global Ocean", anomaly=False,
-                                standardise=False, output=True,
-                                output_dir=outputdir)
-
-        hovmoller_lev_time_plot(data=data, region="Global Ocean",
-                                anomaly=True, standardise=False,
-                                anomaly_ref='Tmean', output=True,
-                                output_dir=outputdir)
-
-        hovmoller_lev_time_plot(data=data, region="Global Ocean",
-                                anomaly=True, standardise=True,
-                                anomaly_ref='Tmean', output=True,
-                                output_dir=outputdir)
+        if predefined_regions:
+            predefined_regions = ocean3d_config["predefined_regions"] ### add fix if not present
+            for predefined_region in predefined_regions:
+                logger.info("Analysing predefined regions")
+                logger.debug("predefined_region: %s", predefined_region)
+                ocean3d_diags(data,
+                              region=predefined_region,
+                              output_dir=outputdir, loglevel=loglevel)
     except AttributeError:
-        print("NoDataError: so or ocpt not found in the Dataset.")
-        print("Not plotting hovmoller_lev_time_plot")
-
-    try:
-        time_series_multilevs(data=data, region='Global Ocean', anomaly=False,
-                              standardise=False, anomaly_ref="FullValue",
-                              customise_level=False, levels=list, output=True,
-                              output_dir=outputdir)
-
-        time_series_multilevs(data=data, region='Global Ocean', anomaly=True,
-                              standardise=False, anomaly_ref="t0",
-                              customise_level=False, levels=list, output=True,
-                              output_dir=outputdir)
-    except AttributeError:
-        print("NoDataError: so or ocpt not found in the Dataset.")
-        print("Not plotting time_series_multilevs")
-    except ValueError:
-        print("ValueError: No levels provided")
-        print("Not plotting time_series_multilevs")
-
-    try:
-        multilevel_t_s_trend_plot(data=data, region='Global Ocean',
-                                  customise_level=False,
-                                  levels=None, output=True,
-                                  output_dir=outputdir)
-    except AttributeError:
-        print("NoDataError: so or ocpt not found in the Dataset.")
-        print("Not plotting multilevel_t_s_trend_plot")
-
-    try:
-        plot_stratification(data, region="Labrador Sea", time="February",
-                            output=True, output_dir=outputdir)
-        plot_stratification(data, region="Labrador Sea", time="DJF",
-                            output=True, output_dir=outputdir)
-    except NoObservationError:
-        print("NoObservationError: No observation available")
-        print("Not plotting plot_stratification")
-    except AttributeError:
-        print("NoDataError: so or ocpt not found in the Dataset.")
-        print("Not plotting plot_stratification")
-
-    try:
-        plot_spatial_mld_clim(data, region="labrador_gin_seas", time="Mar",
-                            overlap=True, output=True, output_dir=outputdir)
-        plot_spatial_mld_clim(data, region="labrador_gin_seas", time="FMA",
-                            overlap=True, output=True, output_dir=outputdir)
-    except NoObservationError:
-        print("NoObservationError: No observation available")
-        print("Not plotting plot_spatial_mld_clim")
-    except AttributeError:
-        print("NoDataError: so or ocpt not found in the Dataset.")
-        print("Not plotting plot_spatial_mld_clim")
+        logger.error("NoDataError: so or ocpt not found in the Dataset.")
+        logger.critical("Not producting ocean diagnostics for predefined regions.")
+    except Exception as e:
+        logger.error(f"Error: {e}, not producting ocean diagnostics for predefined regions.")
+        logger.critical("This could a bug, please report it to the developers.")