+++ diagnostics/tropical_cyclones/cli/tropical_cyclones_cli.py
@@ -6,8 +6,6 @@
 import argparse
 import sys
 
-#sys.path.insert(0, '../')
-
 from tropical_cyclones import TCs
 from aqua.util import load_yaml, get_arg
 from aqua.logger import log_configure
@@ -29,33 +27,30 @@ def parse_arguments(args):
                         required=False)
     parser.add_argument('--source2d', type=str, help='2d source name',
                         required=False)
-    
     parser.add_argument('--source3d', type=str, help='3d source name',
                         required=False)
-    
     parser.add_argument('--outputdir', type=str, help='full res output directory',
                         required=False)
 
     return parser.parse_args(args)
 
 if __name__ == '__main__':
-    
+
     print('Running tropical cyclones diagnostic...')
-    
+
     args = parse_arguments(sys.argv[1:])
 
     # Read configuration file
 
     file = get_arg(args, 'config', 'config_tcs.yaml')
-    print('Reading tcs configuration yaml file.')
+    print('Reading tcs configuration yaml file %s', file)
     config = load_yaml(file)
 
-    # logger setup (via config or command line)
-    
+    # logger setup (via config or clommand line)
+
     loglevel = get_arg(args, 'loglevel', config['setup']['loglevel'])
     logger = log_configure(log_level=loglevel, log_name='TC')
 
-    
     # override config args in case they are passed from command line
 
     model = get_arg(args, 'model', config['dataset']['model'])
@@ -63,15 +58,15 @@ if __name__ == '__main__':
     source2d = get_arg(args, 'source2d', config['dataset']['source2d'])
     source3d = get_arg(args, 'source3d', config['dataset']['source3d'])
     paths = get_arg(args, 'outputdir', config['paths']['fulldir'])
-    
+
     # initialise tropical class with streaming options
-    tropical = TCs(tdict=config, streaming=True, 
-                    model=model,exp=exp, source2d=source2d, source3d=source3d,
-                    stream_step=config['stream']['streamstep'], 
-                    stream_unit="days", 
-                    stream_startdate=config['time']['startdate'], 
-                    paths=paths,
-                    loglevel = loglevel,
-                    nproc=1)
-    
+    tropical = TCs(tdict=config, streaming=True,
+                   model=model, exp=exp, source2d=source2d, source3d=source3d,
+                   stream_step=config['stream']['streamstep'],
+                   stream_unit="days",
+                   stream_startdate=config['time']['startdate'],
+                   paths=paths,
+                   loglevel=loglevel,
+                   nproc=1)
+
     tropical.loop_streaming(config)