+++ diagnostics/tropical_cyclones/tropical_cyclones/__init__.py
@@ -8,5 +8,7 @@ from .detect_nodes import DetectNodes
 from .stitch_nodes import StitchNodes
 from .plots.plotting_hist import plot_hist_cat, plot_press_wind
 
+__version__ = "0.1.0"
+
 # This specifies which methods are exported publicly, used by "from tropical cyclones import *"
 __all__ = ["TCs", "DetectNodes", "StitchNodes", "plot_hist_cat", "plot_pres_wind"]