+++ CHANGELOG.md
@@ -7,6 +7,32 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 Unreleased is the current development version.
 
+- Fix loading of singularity mounting /projappl (#612)
+- CLI wrapper parser (#599)
+- Refactoring of streaming emulator (#593)
+- Radiation CLI and diagnostic refinement (#537)
+- Ocean3D CLI and diagnostic refinement (#578)
+- AtmGlobalMean CLI and diagnostic refinement (#587)
+- Tropical cyclones CLI refinements and TC module (#568)
+- Removing OPA, OPAgenerator and related tests from the AQUA (Remove OPA from AQUA #586)
+- Renaming the experiments according to the DE340 AQUA syntax (Including dev-control-1990 in the source and rename the experiment according to DE340 scheme #556, #614)
+- Allow specifying grid in catalogue metadata (#520)
+- Teleconnections diagnostic improvements (#571, #574, #576, #581, #592)
+
+## [v0.4]
+
+Main changes are:
+1. Update to all the diagnostics CLI
+2. Refactor of the regridder so that `regrid.yaml`` is grid-based and not experiment-based
+3. Xarray access to FDB sources
+4. Refactor of the fixer so that merge/replace/default options are available
+5. Remove of the `aqua` environment in favour of the `aqua_common` one. 
+
+Complete list:
+- Introduced color scheme for aqua logging (#567)
+- CLI for sea diagnostic (#549)
+- Add CLI for SSH diagnostic and some bug fixes (#540)
+- Fix SSH diagnostic to be compatible with lates AQUA version (#538) 
 - Helper function to identify vertical coordinates in a dataset (#552)
 - Orography for tempest extremes TCs detection and update TCs CLI (Orography threshold included and CLI update #404)
 - Improvement of performance indices CLI (Update of ECmean CLI #528)
@@ -15,7 +41,7 @@ Unreleased is the current development version.
 - Custom exceptions for AQUA (#518)
 - Speed up of the `retrieve_plain` method (#524)
 - Update documention for adding new data and setting up the container (Increase documentation coverage #519)
-- CLI wrapper for the state-of-the-art diagnostics analysis (#517, #527, #525, #530, #534, #548)
+- CLI wrapper for the state-of-the-art diagnostics analysis (#517, #527, #525, #530, #534, #536, #539, #548, #549, #559)
 - Refactor the regrid.yaml as grid-based instead of experiment-based (#291)
 - aqua_common environment simplified and updated (#498)
 - Update available variables in FDB catalogues on lumi (#514)
@@ -139,7 +165,8 @@ This is mostly built on the `AQUA` `Reader` class which support for climate mode
 This is the AQUA pre-release to be sent to internal reviewers. 
 Documentations is completed and notebooks are working.
 
-[unreleased]: https://github.com/oloapinivad/AQUA/compare/v0.3...HEAD
+[unreleased]: https://github.com/oloapinivad/AQUA/compare/v0.4...HEAD
+[v0.4]: https://github.com/oloapinivad/AQUA/compare/v0.3...v0.4
 [v0.3]: https://github.com/oloapinivad/AQUA/compare/v0.2.1...v0.3
 [v0.2.1]: https://github.com/oloapinivad/AQUA/compare/v0.2...v0.2.1
 [v0.2]: https://github.com/oloapinivad/AQUA/compare/v0.2-beta...v0.2